// apps/api/src/auth/interfaces/user.google.type.ts
export interface GoogleUser {
  googleId?: string;
  email: string;
  firstName?: string;
  lastName?: string;
  picture?: string;
  accessToken: string;
  refreshToken: string;
}

export interface GoogleLoginInput {
  googleId?: string;
  email: string;
  firstName?: string;
  lastName?: string;
  picture?: string;
}